requirements.txt
langchain
langchain-groq
langchain-ollama
pydantic
python-dotenv
